Clarify that a Qualifying Audit is needed.
The current language suggests that a Qualified Audit is needed. According to
ISAE 3000 and the WebTrust International Reporting under ISAE3000, the term
Qualified has a different meaning within the context of Audits. This commit
clarifies that a Qualifying Audit is needed, according to the requirements
stated in https://aka.ms/auditreqs.