Accessibility improvements
Not everyone is blessed with sight. Some may have to read articles via screen readers. This PR fixes problems that render parts of this text inaccessible via text-to-speech. Incorrectly using a comma, when a full stop is needed, might not impact visual reading, but it has an impact on text-to-speech software. In one instance, this PR rewrites a sentence, so that it is perceived well on audio.

This PR also makes punctuations consistent, i.e. quotation marks, apostrophes, and hyphens.

Finally, this PR fixes three grammatical mistakes. They are jarring on both audio and screen.