Update zero-trust-identity-device-access-policies-common.md
Updating five links to restored content for level-based guiance for Android, after I restored that content in the Inune Docs library. 

Also updated each links to Intune content:   In February, the repository changed from /mem/intune/ to /intune/intune-service/   

Redirects are in place for the old path, but it makes sense to fix all links in this article while updating links to the restored content.